EE: SD and SW instructions

+++ src/memory.cpp
@@ -27,7 +27,7 @@ void Memory::WriteVirtualMemory32(u32 address, u32 dword) {
 u64 Memory::ReadVirtualMemory64(u32 address) {
 	u32 lo = ReadVirtualMemory32(address);
 	u32 hi = ReadVirtualMemory32(address + 1);
-	return (hi << 32) | lo;
+	return ((u64)hi << 32) | lo;
 }
 
 void Memory::WriteVirtualMemory64(u32 address, u64 qword) {
